server: prevent netcode events from triggering shard crashes
When an uncaught exception occurs while handling a packet, this
change logs information and kicks the player responsible from the game.
This helps avoid players being in an invalid state while also
bypassing the need for crashing the whole server.

Note that if other logic prompts an uncaught exception (server ticks, bot code, leaderboard, etc.)
the shard will still die as we assume that the server must be in a bad state.