player: fix kick not closing socket, hack to avoid crash on production
When a player is kicked (afk, no lives, kicked by administrator, etc.), their socket is not
closed. This leaves connections lying open on the server, tying up resources and wasting
bandwidth.

This also includes a small change to avoid a crash when a player is "killed" by 2 seperate
objects in a single tick resulting in them running out of lives. This will be removed
when MongoDB is merged into master, as this race condition will no longer be present.