asteroid: re-introduce patch for sector flooding
The last change was not as efficient as I liked (added a bool for every asteroid, a minimum
of mapSz^2 * minAsteroidsPerSector (8) bools for simple natural checking).

This change has the asteroid class count the number of asteroids per sector, refusing to respawn
if we are over the minimum -- achieving the same effect, but with a smaller memory footprint.

This also optimizes Asteroid::tick() to use the value instead of Object.keys()