netcode: enable support for unix sockets, resolves #208
On https://torn.space, we use an NGINX reverse proxy that is responsible
for passing through the socket.io connections. Currently, we just handle this by passing through to localhost.

However, UNIX domain sockets enable us to skip a lot of the overhead with the network stack while still maintaining a transparent TCP interface. This hopefully reduces additional latency associated with the reverse proxying.