netcode: don't crash if 'error' is emitted
Apparently the default behavior is to just crash if this is the
case. And of course, any client can just emit this message, so
frankly I'm not sure what the point of this mechanism is. We already
have exceptions on the sender side and the receiver side can raise exceptions on their end.

Anyhow, this should fix up some of the constant crashes and improve logging
slightly.